2009-05-03  Jeffrey Stedfast  <fejj@novell.com>
	* runtime.h (InMainThread): check main_thread_inited.

	* runtime.cpp: All flags variables are now 32bit. Split _EX flags
	into a second debug_flags_ex. Also added a new bool var to keep
	track of whether or not main_thread has been initialized since we
	can't necessary compare against NULL.

	* debug.h: Updated.


svn path=/trunk/moon/; revision=133437

+++ plugin/plugin.cpp
@@ -51,7 +51,7 @@
 // Debug NPStreams
 #define nps(x)//x
 
-extern guint64 moonlight_flags;
+extern guint32 moonlight_flags;
 
 /* gleaned from svn log of the moon module, as well as olive/class/{agclr,agmono,System.Silverlight} */
 static const char *moonlight_authors[] = {
